Классифицировал сущности. Добавил следующие модели : Session, Review
