--- src/Core/Domain/Entities/User.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Entities
-{
-	public class User : Entity
-	{
-		public string Username { get; set; }
-	}
-}
