+++ src/Core/Domain/Entities/_Auth/Session.cs
@@ -3,9 +3,11 @@
 
 namespace Domain.Entities._Auth
 {
-	public class Sessions
+	public class Session
 	{
 		public Guid ID { get; set; }
+
+		public string AppName { get; set; }
 
 		public DateTime LastUpdate { get; set; }
 	}
