+++ src/Core/Domain/Entities/_Relationships/BookAuthor.cs
@@ -1,15 +1,16 @@
 ﻿using Domain.Entities._Book;
+using Domain.Entities._Creators;
 
 namespace Domain.Entities._Relationships
 {
     public class BookAuthor
-    {
-        #region Relationships
-        public long BookID { get; set; }
-        public Book Book { get; set; }
+	{
+		#region Relationships
+		public long BookID { get; set; }
+		public Book Book { get; set; }
 
-        public long AuthorID { get; set; }
-        public Author Author { get; set; }
-        #endregion
-    }
+		public long AuthorID { get; set; }
+		public Author Author { get; set; }
+		#endregion
+	}
 }
