+++ src/Core/Domain/Entities/_User/User.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Entities._User
+{
+    public class User : Entity
+    {
+        public string Username { get; set; }
+    }
+}
