+++ src/Core/Domain/Enums/ReviewType.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Enums
+{
+	public enum ReviewType
+	{
+		Positive, Neutral, Negative
+	}
+}
